Fix issue with deriving relative paths in Ruby 2.2
When determining what files to create symlinks for, there is a 'sub'
statement that peels off the root of a filepath to derive a relative
path. This works in Ruby 2.1, because both of the paths happen to be
lowercase, but a fix in Ruby 2.2 has broken it.

Namely, if you had a file called "Filename" (note the case) this
behavior changed:
 # (Ruby 2.1)
Dir['filename'] #=> ['filename']

 # (Ruby 2.2)
Dir['filename'] #=> ['Filename']

Now, we ensure that the path for "pref_sources/RubyMine" has the correct
case, AND change the path sub code to be case insensitive.